Review commits

+++ src/main/java/com/snowflake/kafka/connector/SnowflakeSinkConnectorConfig.java
@@ -498,7 +498,7 @@ public class SnowflakeSinkConnectorConfig {
             ERRORS_DEAD_LETTER_QUEUE_TOPIC_NAME_DISPLAY)
         .define(
             SNOWFLAKE_CLEANER_MAX_RETRIES_CONFIG,
-            Type.INT,
+            Type.LONG,
             SNOWFLAKE_CLEANER_MAX_RETRIES_DEFAULT,
             Importance.LOW,
             "Number of times to retry a failed cleaner operation. Defaults to unbounded retries.",
